refactor: Consolidate test helpers and add missing newline
- Add newline at end of button.test.ts
- Consolidate test helper functions in test/utils/test-helpers.ts
- Re-export helpers from tests/setup.ts for backward compatibility
- Add createTestElement, cleanupTestElement, waitForCustomElement
- Add getShadowContent as alias for getShadowElement (deprecated)
